+++ src/commands/registerCommand.ts
@@ -1,7 +1,10 @@
 import * as vscode from "vscode";
 import { type RegionStore } from "../state/RegionStore";
+import { type FullTreeViewProvider } from "../treeView/fullTreeView/FullTreeViewProvider";
 import { goToFullTreeItemCommand } from "../treeView/fullTreeView/goToFullTreeItem";
 import { goToRegionTreeItemCommand } from "../treeView/regionTreeView/goToRegionTreeItem";
+import { type RegionTreeViewProvider } from "../treeView/regionTreeView/RegionTreeViewProvider";
+import { allExpandAllCommands } from "./expandAndCollapseAll";
 import { goToNextRegionCommand } from "./goToNextRegion";
 import { goToPreviousRegionCommand } from "./goToPreviousRegion";
 import { goToRegionBoundaryCommand } from "./goToRegionBoundary";
@@ -14,23 +17,38 @@ type RegionHelperExtensionId = "regionHelper";
 
 type RegionHelperCommandId = `${RegionHelperExtensionId}.${string}`;
 
-export type RegionHelperCommand = {
+export type RegionHelperClosuredParams = {
+  regionStore: RegionStore;
+  regionTreeViewProvider: RegionTreeViewProvider;
+  fullTreeViewProvider: FullTreeViewProvider;
+};
+
+/** A command that needs access to RegionHelperClosuredParams. */
+export type RegionHelperClosuredCommand = {
+  id: RegionHelperCommandId;
+  callback: (regionHelperParams: RegionHelperClosuredParams) => void;
+  needsRegionHelperParams: true;
+};
+
+/** A command that doesn't need access to RegionHelperClosuredParams. */
+export type RegionHelperNonClosuredCommand = {
   id: RegionHelperCommandId;
   callback: Parameters<typeof vscode.commands.registerCommand>[1];
-  // TODO: better type safety for this (make sure needsRegionStore is true if and only if callback needs it)
-  needsRegionStore: boolean;
+  needsRegionHelperParams: false;
 };
+
+type RegionHelperCommand = RegionHelperClosuredCommand | RegionHelperNonClosuredCommand;
 
 export function registerAllCommands(
   subscriptions: vscode.Disposable[],
-  regionStore: RegionStore
+  regionHelperParams: RegionHelperClosuredParams
 ): void {
   for (const command of commandsToRegister) {
-    registerRegionHelperCommand(command, subscriptions, regionStore);
+    registerRegionHelperCommand(command, subscriptions, regionHelperParams);
   }
 }
 
-const commandsToRegister = [
+const commandsToRegister: RegionHelperCommand[] = [
   goToRegionTreeItemCommand,
   goToFullTreeItemCommand,
   goToRegionBoundaryCommand,
@@ -40,24 +58,18 @@ const commandsToRegister = [
   goToPreviousRegionCommand,
   ...allRegionsViewConfigCommands,
   ...allFullOutlineViewConfigCommands,
+  ...allExpandAllCommands,
 ];
 
 function registerRegionHelperCommand(
   command: RegionHelperCommand,
   subscriptions: vscode.Disposable[],
-  regionStore: RegionStore
+  regionHelperParams: RegionHelperClosuredParams
 ): void {
-  const { id, callback, needsRegionStore } = command;
+  const { id, callback, needsRegionHelperParams } = command;
   const commandDisposable = vscode.commands.registerCommand(
     id,
-    needsRegionStore ? makeRegionStoreCallback(regionStore, callback) : callback
+    needsRegionHelperParams ? (): void => callback(regionHelperParams) : callback
   );
   subscriptions.push(commandDisposable);
-}
-
-function makeRegionStoreCallback(
-  regionStore: RegionStore,
-  callback: (regionStore: RegionStore) => void
-): () => void {
-  return () => callback(regionStore);
 }